# openapi.yaml
# modified the specification to include error responses,
# page and limit query parameters are now listed
# messages/send is now post route with modified parameters
openapi: 3.0.0
info:
  title: Messages API
  version: 1.0.0
paths:
  /messages:
    get:
      summary: Get Messages
      description: Retrieves a list of messages.
      parameters:
        - name: status
          in: query
          description: Filter messages by status
          required: false
          schema:
            type: "string"
            enum:
              - "sent"
              - "read"
              - "draft"
        - name: page
          in: query
          description: Page number for pagination
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Number of items per page
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 10
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  messages:
                    type: "array"
                    items:
                      type: "object"
                      properties:
                        uuid: 
                          type: "string"
                        text:
                          type: "string"
                        status:
                          type: "string"
                          enum:
                            - "sent"
                            - "read"
                            - "draft"
              example:
                messages:
                  - text: "Hello, World!"
                    status: "read"
                  - text: "How are you today?"
                    status: "sent"
                  - text: "Thanks, I'm fine. How about you?"
                    status: "sent"
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/ServerError'
  /messages/send:
    post:
      summary: Send a Message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - text
              properties:
                text:
                  type: string
                  maxLength: 2000
                  minLength: 1
      responses:
        '204':
          description: Successfully sent
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/ServerError'

components:
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string